+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning].
 
 ## [Unreleased]
 
+- [PR-35](https://github.com/itk-dev/ai-screening/pull/35)
+  Added frontpage
+  Changed project view
+  Added twig function for project status display
 - [PR-33](https://github.com/itk-dev/ai-screening/pull/33)
   OIDC tweaks
 - [PR-32](https://github.com/itk-dev/ai-screening/pull/32)